pi_server/requirements.txt
click==8.1.3
dlib==19.9.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.3.0
pyserial==3.5
